+++ auction/src/main/java/com/smore/auction/infrastructure/websocket/manager/AuctionSessionManager.java
@@ -1,4 +1,4 @@
-package com.smore.auction.infrastructure.websocket.handler;
+package com.smore.auction.infrastructure.websocket.manager;
 
 public interface AuctionSessionManager {
     void handleSubscribe(String sessionId, Long userId, String auctionId);
